Hoist tightly coupled `Suit` and `Strain` to crate root

+++ tests/deal.rs
@@ -1,4 +1,5 @@
 use core::num::Wrapping;
+use dds_bridge::Suit;
 use dds_bridge::deal::*;
 
 const _: () = {
@@ -64,18 +65,9 @@ fn test_seat_arithmetics() {
 
     (0..4).for_each(|x| {
         (0..4).for_each(|y| {
-            assert_eq!(
-                SEATS[usize::from(x)] + Wrapping(y),
-                SEATS[usize::from(y)] + Wrapping(x)
-            );
-            assert_eq!(
-                SEATS[usize::from(x)] + Wrapping(y),
-                SEATS[usize::from((x + y) & 3)]
-            );
-            assert_eq!(
-                SEATS[usize::from(x)] - Wrapping(y),
-                SEATS[usize::from((x + 4 - y) & 3)]
-            );
+            assert_eq!(SEATS[x] + Wrapping(y), SEATS[y] + Wrapping(x));
+            assert_eq!(SEATS[x] + Wrapping(y), SEATS[(x + y) & 3]);
+            assert_eq!(SEATS[x] - Wrapping(y), SEATS[(x + 4 - y) & 3]);
         });
     });
 }
